requirements.txt
astroid==1.5.3
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bleach==1.5.0
brewer2mpl==1.4.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
dlib==19.7.0
entrypoints==0.2.3
enum34==1.1.6
face-recognition==1.0.0
face-recognition-models==0.3.0
Flask==0.12.2
funcsigs==1.0.2
functools32==3.2.3.post2
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipdb==0.10.3
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
Keras==2.0.8
lazy-object-proxy==1.3.1
mccabe==0.6.1
mistune==0.7.4
mock==2.0.0
networkx==2.0
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pandocfilters==1.4.2
pathlib2==2.3.0
patsy==0.4.1
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.5.2
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
scandir==1.5
six==1.10.0
subprocess32==3.2.7
tensorflow==1.3.0
tensorflow-tensorboard==0.1.5
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
Unipath==1.1
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.3
wrapt==1.10.11
opencv-python==3.3.0.10
gunicorn==19.7.1